+++ supertokens_python/__init__.py
@@ -16,7 +16,7 @@ from typing import Any, Callable, Dict, List, Optional
 
 from typing_extensions import Literal
 
-from supertokens_python import async_to_sync_wrapper
+from supertokens_python.async_to_sync.handler import ConcreteAsyncHandler
 from supertokens_python.framework.request import BaseRequest
 from supertokens_python.types import RecipeUserId
 
@@ -37,7 +37,7 @@ def init(
     mode: Optional[Literal["asgi", "wsgi"]] = None,
     telemetry: Optional[bool] = None,
     debug: Optional[bool] = None,
-    async_handler: Optional[async_to_sync_wrapper.ConcreteAsyncHandler] = None,
+    async_handler: Optional[ConcreteAsyncHandler] = None,
 ):
     return Supertokens.init(
         app_info=app_info,
