+++ supertokens_python/recipe/emailverification/syncio/__init__.py
@@ -14,7 +14,7 @@
 
 from typing import Any, Dict, Optional, Union
 
-from supertokens_python.async_to_sync_wrapper import sync
+from supertokens_python.async_to_sync.base import sync
 from supertokens_python.recipe.emailverification.types import EmailTemplateVars
 from supertokens_python.types import RecipeUserId
