+++ supertokens_python/recipe/multitenancy/syncio/__init__.py
@@ -14,7 +14,7 @@
 from __future__ import annotations
 from typing import Any, Dict, Optional, TYPE_CHECKING
 
-from supertokens_python.async_to_sync_wrapper import sync
+from supertokens_python.async_to_sync.base import sync
 from supertokens_python.recipe.multitenancy.interfaces import TenantConfigCreateOrUpdate
 from supertokens_python.types import RecipeUserId
