+++ supertokens_python/recipe/openid/syncio/__init__.py
@@ -13,7 +13,7 @@
 # under the License.
 from typing import Any, Dict, Union, Optional
 
-from supertokens_python.async_to_sync_wrapper import sync
+from supertokens_python.async_to_sync.base import sync
 from supertokens_python.recipe.openid import asyncio
 from supertokens_python.recipe.openid.interfaces import (
     GetOpenIdDiscoveryConfigurationResult,
