+++ supertokens_python/recipe/session/interfaces.py
@@ -27,7 +27,7 @@ from typing import (
 )
 from typing_extensions import TypedDict
 
-from supertokens_python.async_to_sync_wrapper import sync
+from supertokens_python.async_to_sync.base import sync
 from supertokens_python.types import (
     APIResponse,
     GeneralErrorResponse,
