+++ supertokens_python/recipe/userroles/syncio/__init__.py
@@ -14,7 +14,7 @@
 
 from typing import Any, Dict, List, Union
 
-from supertokens_python.async_to_sync_wrapper import sync
+from supertokens_python.async_to_sync.base import sync
 from supertokens_python.recipe.userroles.interfaces import (
     AddRoleToUserOkResult,
     CreateNewRoleOrAddPermissionsOkResult,
